feat: cities crud

+++ src/Controller/UserManagement.php
@@ -33,6 +33,9 @@ class UserManagement extends AbstractController
     #[Route(name: 'users', path: '/users', methods: 'GET')]
     public function list(Request $request): Response
     {
+        if(!$request->getSession()->has('login')){
+            return $this->redirectToRoute('login');
+        }
         $page = $request->query->get('page', 1);
 
         $entities = $this->model->all(limit: 10, page: $page - 1);
@@ -45,6 +48,9 @@ class UserManagement extends AbstractController
     #[Route('/user/new', methods: [ 'GET', 'POST' ])]
     public function create(Request $request): Response
     {
+        if(!$request->getSession()->has('login')){
+            return $this->redirectToRoute('login');
+        }
         $user = new User();
         $form = $this->createForm(UserForm::class, $user);
         $form->handleRequest($request);
@@ -64,6 +70,9 @@ class UserManagement extends AbstractController
     #[Route('/user/{id}/edit', methods: [ 'GET', 'POST' ])]
     public function edit(Request $request, int $id): Response
     {
+        if(!$request->getSession()->has('login')){
+            return $this->redirectToRoute('login');
+        }
         $user = $this->ifEntity($id);
         $form = $this->createForm(UserForm::class, $user);
         $form->handleRequest($request);
@@ -84,6 +93,9 @@ class UserManagement extends AbstractController
     #[Route('/user/{id}/delete', methods: 'GET')]
     public function delete(Request $request, int $id): Response
     {
+        if(!$request->getSession()->has('login')){
+            return $this->redirectToRoute('login');
+        }
         $user = $this->ifEntity($id);
         if ($request->query->has('confirmation')) {
             $this->model->delete($user);
@@ -93,8 +105,11 @@ class UserManagement extends AbstractController
     }
 
     #[Route('/user/{id}', methods: 'GET')]
-    public function details(int $id): Response
+    public function details(Request $request, int $id): Response
     {
+        if(!$request->getSession()->has('login')){
+            return $this->redirectToRoute('login');
+        }
         $entity = $this->ifEntity($id);
         return $this->render('user.html.twig', ['user' => $entity]);
     }
